+++ udata/commands/fixtures.py
@@ -1,5 +1,6 @@
 import json
 import logging
+import pathlib
 
 import click
 import requests
@@ -27,15 +28,65 @@ COMMUNITY_RES_URL = "/api/1/datasets/community_resources"
 DISCUSSION_URL = "/api/1/discussions"
 
 
-DEFAULT_FIXTURE_FILE = (
+DEFAULT_FIXTURE_FILE: str = (
     "https://raw.githubusercontent.com/opendatateam/udata-fixtures/main/results.json"  # noqa
 )
+
+DEFAULT_FIXTURES_RESULTS_FILENAME: str = "results.json"
+
+UNWANTED_KEYS: dict[str, list[str]] = {
+    "dataset": [
+        "uri",
+        "page",
+        "last_update",
+        "last_modified",
+        "license",
+        "badges",
+        "spatial",
+        "quality",
+    ],
+    "resource": ["latest", "preview_url", "last_modified"],
+    "organization": ["members", "page", "uri", "logo_thumbnail"],
+    "reuse": ["datasets", "image_thumbnail", "page", "uri", "organization", "owner"],
+    "community": [
+        "dataset",
+        "organization",
+        "owner",
+        "latest",
+        "last_modified",
+        "preview_url",
+    ],
+    "discussion": ["subject", "user", "url", "class"],
+    "message": ["posted_by"],
+}
+
+
+def remove_unwanted_keys(obj: dict, filter_type: str) -> dict:
+    """Remove UNWANTED_KEYS from a dict."""
+    for unwanted_key in UNWANTED_KEYS[filter_type]:
+        if unwanted_key in obj:
+            del obj[unwanted_key]
+    fix_dates(obj)
+    return obj
+
+
+def fix_dates(obj: dict) -> dict:
+    """Fix dates from the fixtures so they can be safely reloaded later on."""
+    if "internal" not in obj:
+        return obj
+    obj["created_at_internal"] = obj["internal"]["created_at_internal"]
+    obj["last_modified_internal"] = obj["internal"]["last_modified_internal"]
+    del obj["internal"]
+    del obj["created_at"]
+    return obj
 
 
 @cli.command()
 @click.argument("data-source")
-def generate_fixtures_file(data_source):
+@click.argument("results-filename", default=DEFAULT_FIXTURES_RESULTS_FILENAME)
+def generate_fixtures_file(data_source: str, results_filename: str) -> None:
     """Build sample fixture file based on datasets slugs list (users, datasets, reuses)."""
+    results_file = pathlib.Path(results_filename)
     datasets_slugs = current_app.config["FIXTURE_DATASET_SLUGS"]
     json_result = []
 
@@ -44,31 +95,13 @@ def generate_fixtures_file(data_source):
             json_fixture = {}
 
             json_dataset = requests.get(f"{data_source}{DATASET_URL}/{slug}/").json()
-            del json_dataset["uri"]
-            del json_dataset["page"]
-            del json_dataset["last_update"]
-            del json_dataset["last_modified"]
-            del json_dataset["license"]
-            del json_dataset["badges"]
-            del json_dataset["spatial"]
-            del json_dataset["quality"]
-            json_dataset["created_at_internal"] = json_dataset.pop("created_at")
             json_resources = json_dataset.pop("resources")
-            for res in json_resources:
-                del res["latest"]
-                del res["preview_url"]
-                del res["last_modified"]
-                res["created_at_internal"] = res.pop("created_at")
             if json_dataset["organization"] is None:
                 json_owner = json_dataset.pop("owner")
                 json_dataset["owner"] = json_owner["id"]
             else:
                 json_org = json_dataset.pop("organization")
                 json_org = requests.get(f"{data_source}{ORG_URL}/{json_org['id']}/").json()
-                del json_org["members"]
-                del json_org["page"]
-                del json_org["uri"]
-                del json_org["logo_thumbnail"]
                 json_fixture["organization"] = json_org
             json_fixture["resources"] = json_resources
             json_fixture["dataset"] = json_dataset
@@ -76,49 +109,28 @@ def generate_fixtures_file(data_source):
             json_reuses = requests.get(
                 f"{data_source}{REUSE_URL}/?dataset={json_dataset['id']}"
             ).json()["data"]
-            for reuse in json_reuses:
-                del reuse["datasets"]
-                del reuse["image_thumbnail"]
-                del reuse["page"]
-                del reuse["uri"]
-                del reuse["organization"]
-                del reuse["owner"]
             json_fixture["reuses"] = json_reuses
 
             json_community = requests.get(
                 f"{data_source}{COMMUNITY_RES_URL}/?dataset={json_dataset['id']}"
             ).json()["data"]
-            for com in json_community:
-                del com["dataset"]
-                del com["organization"]
-                del com["owner"]
-                del com["latest"]
-                del com["last_modified"]
-                del com["preview_url"]
-                com["created_at_internal"] = com.pop("created_at")
             json_fixture["community_resources"] = json_community
 
             json_discussion = requests.get(
                 f"{data_source}{DISCUSSION_URL}/?for={json_dataset['id']}"
             ).json()["data"]
-            for discussion in json_discussion:
-                del discussion["subject"]
-                del discussion["user"]
-                del discussion["url"]
-                del discussion["class"]
-                for message in discussion["discussion"]:
-                    del message["posted_by"]
             json_fixture["discussions"] = json_discussion
 
             json_result.append(json_fixture)
 
-    with open("results.json", "w") as f:
-        json.dump(json_result, f)
+    with results_file.open("w") as f:
+        json.dump(json_result, f, indent=2)
+        print(f"Fixtures saved to file {results_filename}")
 
 
 @cli.command()
 @click.argument("source", default=DEFAULT_FIXTURE_FILE)
-def generate_fixtures(source):
+def import_fixtures(source):
     """Build sample fixture data (users, datasets, reuses) from local or remote file."""
     if source.startswith("http"):
         json_fixtures = requests.get(source).json()
@@ -129,24 +141,32 @@ def generate_fixtures(source):
     with click.progressbar(json_fixtures) as bar:
         for fixture in bar:
             user = UserFactory()
+            dataset = fixture["dataset"]
+            dataset = remove_unwanted_keys(dataset, "dataset")
             if not fixture["organization"]:
-                dataset = DatasetFactory(**fixture["dataset"], owner=user)
+                dataset = DatasetFactory(**dataset, owner=user)
             else:
                 org = Organization.objects(id=fixture["organization"]["id"]).first()
                 if not org:
-                    org = OrganizationFactory(
-                        **fixture["organization"], members=[Member(user=user)]
-                    )
-                dataset = DatasetFactory(**fixture["dataset"], organization=org)
+                    organization = fixture["organization"]
+                    organization = remove_unwanted_keys(organization, "organization")
+                    org = OrganizationFactory(**organization, members=[Member(user=user)])
+                dataset = DatasetFactory(**dataset, organization=org)
             for resource in fixture["resources"]:
+                resource = remove_unwanted_keys(resource, "resource")
                 res = ResourceFactory(**resource)
                 dataset.add_resource(res)
             for reuse in fixture["reuses"]:
+                reuse = remove_unwanted_keys(reuse, "reuse")
                 ReuseFactory(**reuse, datasets=[dataset], owner=user)
             for community in fixture["community_resources"]:
+                community = remove_unwanted_keys(community, "community")
                 CommunityResourceFactory(**community, dataset=dataset, owner=user)
             for discussion in fixture["discussions"]:
+                discussion = remove_unwanted_keys(discussion, "discussion")
                 messages = discussion.pop("discussion")
+                for message in messages:
+                    message = remove_unwanted_keys(message, "message")
                 DiscussionFactory(
                     **discussion,
                     subject=dataset,